Separate ROS node and TUI implemented. IK implemented.
